Refactor path names
